[Code.org only] Prepare for npm publishing
We'll be publishing this package on npm as `@code-dot-org/p5.play` to avoid a collision if @molleindustria wishes to publish the main project someday. In addition, we will only be publishing versions with `-cdo` [prerelease tags](https://docs.npmjs.com/misc/semver#prerelease-tags) to avoid version numbering collisions with the main project.

I've updated the package.json to point to our own repositories, etc, to credit known contributors, and to whiteliest the p5.js and p5.play.js files so that a minimum amount of content is actually pushed to npm.  To publish a new version one will run `npm version <new version id>` where the new id must take the form `#.#.#-cdo*`.  This will run all tests to verify the state of the library, apply the new version number, push new tags to GitHub and new content to npm.