Use scaleX and scaleY helpers when setting parent transform
Doesn't make much difference in vanilla p5.play which only permits uniform scale, but if nonuniform scale is supported (as in Game Lab) this allows the collider to take the sprite's nonuniform scale into account.

There are other problems with nonuniform scale already (like how it's handled when there are also rotations involved) but this gives correct behavior in more cases than before.